Add a method `set_pixels_raw`
In some cases sending data this way is significantly faster than
taking iterators and converting the values on the fly (like
`set_pixels` does). In my case the difference is ~150ms vs ~40ms.
This method provides a simple way to prepare the pixel data in an
arbitrary way and then send it to the display as fast as possible.